requirements.txt
certifi==2025.4.26
contourpy==1.3.2
cycler==0.12.1
elastic-transport==8.17.1
elasticsearch==8.11.1
Faker==37.3.0
fonttools==4.58.0
greenlet==3.2.2
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.41
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
